reorg keyset startup initialization
